desktop/requirements.txt
customtkinter>=5.2.0
pystray>=0.19.5
Pillow>=10.0.0
qrcode>=7.4.0
psutil>=5.9.0
psycopg2-binary>=2.9.9
